requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Authentication
djangorestframework-simplejwt==5.3.0

# AI Services
openai==1.3.7

# Async Processing
celery==5.3.4
redis==5.0.1

# Database
psycopg2-binary==2.9.9

# Environment Variables
python-dotenv==1.0.0

# Server
gunicorn==21.2.0

# Image Processing (optional)
Pillow==10.1.0

# API Documentation (optional)
drf-yasg==1.21.7